src/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# File Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# AWS S3
boto3==1.34.0
botocore==1.34.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# LinkedIn API
linkedin-api==2.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Task Queue
celery==5.3.4
kombu==5.3.4

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Linting
black==23.12.0
flake8==6.1.0
mypy==1.7.1
